# perf(event): get_events performance and Google Calendar integration fixes