--- a/frappe/desk/doctype/event/event.json
+++ b/frappe/desk/doctype/event/event.json
@@ -247,9 +247,9 @@
    "fieldname": "google_calendar_event_id",
    "fieldtype": "Data",
    "label": "Google Calendar Event ID",
+   "length": 320,
    "no_copy": 1,
-   "read_only": 1,
-   "length": 320
+   "read_only": 1
   },
   {
    "default": "0",
@@ -287,16 +287,17 @@
   },
   {
    "fieldname": "google_meet_link",
-   "fieldtype": "Data",
+   "fieldtype": "Small Text",
    "label": "Google Meet Link",
    "no_copy": 1,
    "read_only": 1
   }
  ],
+ "grid_page_length": 50,
  "icon": "fa fa-calendar",
  "idx": 1,
  "links": [],
- "modified": "2023-06-23 10:33:15.685368",
+ "modified": "2025-04-10 13:08:32.540745",
  "modified_by": "Administrator",
  "module": "Desk",
  "name": "Event",
@@ -328,6 +329,7 @@
   }
  ],
  "read_only": 1,
+ "row_format": "Dynamic",
  "sender_field": "sender",
  "sort_field": "modified",
  "sort_order": "DESC",