@gavindsouza 2025-04-16
Will run tests that exist on develop & newly added ones via https://github.com/frappe/frappe/pull/32116 on v15 and re-raise PR with fixes.